fix: update deps in api and client
