+++ packages/api/package.json
@@ -24,16 +24,15 @@
     "cluster:reset": "web3-storage-tools cluster --clean"
   },
   "devDependencies": {
-    "@sentry/webpack-plugin": "^1.16.0",
-    "@types/mocha": "^8.2.2",
-    "@web-std/form-data": "^2.1.0",
+    "@sentry/webpack-plugin": "^1.18.3",
+    "@types/mocha": "^9.0.0",
+    "@web-std/form-data": "^3.0.0",
     "@web3-storage/tools": "^1.0.0",
     "assert": "^2.0.0",
     "buffer": "^6.0.3",
     "dotenv": "^10.0.0",
     "git-rev-sync": "^3.0.1",
     "git-revision-webpack-plugin": "^5.0.0",
-    "ipfs-unixfs-importer": "^9.0.4",
     "mocha": "^9.1.3",
     "npm-run-all": "^4.1.5",
     "playwright-test": "^7.0.1",
@@ -54,13 +53,13 @@
     "@ipld/dag-pb": "^2.0.2",
     "@magic-ext/oauth": "^0.8.0",
     "@magic-sdk/admin": "^1.3.0",
-    "@nftstorage/ipfs-cluster": "^3.3.1",
+    "@nftstorage/ipfs-cluster": "^3.4.3",
     "@web3-storage/db": "^4.0.0",
     "@web3-storage/multipart-parser": "^1.0.0",
-    "cborg": "^1.5.3",
-    "ipfs-car": "^0.5.8",
-    "itty-router": "^2.3.10",
-    "multiformats": "^9.0.4",
+    "cborg": "^1.6.0",
+    "ipfs-car": "^0.6.1",
+    "itty-router": "^2.4.8",
+    "multiformats": "^9.5.8",
     "p-retry": "^4.6.1",
     "protobufjs": "^6.11.2",
     "tweetnacl": "^1.0.3",
