+++ packages/client/package.json
@@ -31,43 +31,43 @@
   },
   "dependencies": {
     "@ipld/car": "^3.1.4",
-    "@web-std/blob": "^2.1.0",
-    "@web-std/fetch": "^2.0.2",
-    "@web-std/file": "^1.1.0",
-    "browser-readablestream-to-it": "^1.0.2",
+    "@web-std/blob": "^3.0.1",
+    "@web-std/fetch": "^3.0.0",
+    "@web-std/file": "^3.0.0",
+    "browser-readablestream-to-it": "^1.0.3",
     "carbites": "^1.0.6",
-    "cborg": "^1.5.3",
-    "files-from-path": "^0.2.1",
-    "ipfs-car": "^0.5.8",
-    "ipns": "^0.15.0",
-    "libp2p-crypto": "^0.20.0",
+    "cborg": "^1.6.0",
+    "files-from-path": "^0.2.3",
+    "ipfs-car": "^0.6.1",
+    "ipns": "^0.16.0",
+    "libp2p-crypto": "^0.21.0",
     "p-retry": "^4.5.0",
     "parse-link-header": "^2.0.0",
     "streaming-iterables": "^6.0.0",
     "uint8arrays": "^3.0.0"
   },
   "devDependencies": {
-    "@ipld/dag-json": "^8.0.0",
-    "@rollup/plugin-commonjs": "^19.0.0",
+    "@ipld/dag-json": "^8.0.7",
+    "@rollup/plugin-commonjs": "^21.0.1",
     "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.0.0",
-    "@types/mocha": "8.2.2",
+    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@types/mocha": "9.0.0",
     "@types/parse-link-header": "^1.0.1",
     "bundlesize": "^0.18.1",
     "cors": "^2.8.5",
     "del-cli": "^4.0.0",
     "hundreds": "0.0.9",
-    "mocha": "8.3.2",
-    "multiformats": "^9.1.2",
+    "mocha": "9.1.3",
+    "multiformats": "^9.5.8",
     "npm-run-all": "^4.1.5",
     "nyc": "15.1.0",
-    "playwright-test": "^7.0.1",
-    "rollup": "2.50.1",
+    "playwright-test": "^7.2.2",
+    "rollup": "2.63.0",
     "rollup-plugin-multi-input": "1.3.1",
     "rollup-plugin-terser": "^7.0.2",
     "smoke": "^3.1.1",
-    "typedoc": "0.20.36",
-    "uvu": "0.5.1"
+    "typedoc": "0.22.10",
+    "uvu": "0.5.3"
   },
   "types": "./dist/src/lib.d.ts",
   "module": "./src/lib.js",
