+++ packages/w3/package.json
@@ -27,10 +27,10 @@
     "conf": "^10.0.1",
     "enquirer": "^2.3.6",
     "hard-rejection": "^2.1.0",
-    "ipfs-car": "^0.5.8",
+    "ipfs-car": "^0.6.1",
     "ora": "^5.4.1",
     "sade": "^1.7.4",
-    "web3.storage": "^3.0.0"
+    "web3.storage": "^3.5.2"
   },
   "devDependencies": {
     "execa": "^5.1.1",
